Pull eval terms into separate file
I caved. Collocation my ass...
